// types.ts
export interface Project {
  id: number;
  title: string;
  category: string;
  image: string;
  alt: string;
}

export interface ExpertiseStep {
  id: number;
  icon: string;
  title: string;
  description: string;
}

export interface VisionItem {
  id: number;
  icon: string;
  title: string;
  subtitle: string;
}